+++ __tests__/core/cluster.ts
@@ -1,4 +1,5 @@
-import { Process } from "./../../src/index";
+import { Process, config } from "./../../src/index";
+import { sleep } from "./../../src/utils/sleep";
 
 const actionhero = new Process();
 let api;
@@ -6,14 +7,12 @@ let api;
 describe("Core: Action Cluster", () => {
   beforeAll(async () => {
     api = await actionhero.start();
-    for (var room in api.config.general.startingChatRooms) {
+    for (var room in config.general.startingChatRooms) {
       try {
         await api.chatRoom.destroy(room);
         await api.chatRoom.add(room);
       } catch (error) {
-        if (
-          !error.toString().match(api.config.errors.connectionRoomExists(room))
-        ) {
+        if (!error.toString().match(config.errors.connectionRoomExists(room))) {
           throw error;
         }
       }
@@ -35,7 +34,7 @@ describe("Core: Action Cluster", () => {
         data[1] = [arg1, arg2];
       };
       await api.redis.doCluster("api.rpcTestMethod", ["arg1", "arg2"]);
-      await api.utils.sleep(100);
+      await sleep(100);
 
       expect(data[1][0]).toEqual("arg1");
       expect(data[1][1]).toEqual("arg2");
@@ -53,7 +52,7 @@ describe("Core: Action Cluster", () => {
         ["arg1", "arg2"],
         client.id
       );
-      await api.utils.sleep(100);
+      await sleep(100);
 
       expect(data[1][0]).toEqual("arg1");
       expect(data[1][1]).toEqual("arg2");
