+++ src/actions/cacheTest.ts
@@ -1,4 +1,4 @@
-import { api, Cache, Action } from "./../index";
+import { cache, Action } from "./../index";
 
 export class CacheTest extends Action {
   constructor() {
@@ -51,10 +51,10 @@ export class CacheTest extends Action {
     const value = params.value;
 
     response.cacheTestResults = {
-      saveResp: await Cache.save(key, value, 5000),
-      sizeResp: await Cache.size(),
-      loadResp: await Cache.load(key),
-      deleteResp: await Cache.destroy(key)
+      saveResp: await cache.save(key, value, 5000),
+      sizeResp: await cache.size(),
+      loadResp: await cache.load(key),
+      deleteResp: await cache.destroy(key)
     };
   }
 }
