+++ src/index.ts
@@ -1,4 +1,5 @@
-// export classes
+// export classes (capitalized)
+export { Api } from "./classes/api";
 export { Process } from "./classes/process";
 export { Initializer } from "./classes/initializer";
 export { Connection } from "./classes/connection";
@@ -8,24 +9,28 @@ export { Server } from "./classes/server";
 export { CLI } from "./classes/cli";
 export { ActionProcessor } from "./classes/actionProcessor";
 
-// export methods
-export { log } from "./classes/log";
-export { localize } from "./classes/i18n";
-export {
-  watchFileAndAct,
-  unWatchAllFiles
-} from "./classes/process/watchFileAndAct";
+// export modules (lower case)
+export { config } from "./modules/config";
+export { log } from "./modules/log";
+export { action } from "./modules/action";
+export { task } from "./modules/task";
+export { cache } from "./modules/cache";
+export { chatRoom } from "./modules/chatRoom";
+export { redis } from "./modules/redis";
+export { i18n } from "./modules/i18n";
+export { route } from "./modules/route";
+export { specHelper } from "./modules/specHelper";
 
-// export objects
-export { config } from "./classes/config";
-export { Cache } from "./classes/cache";
-
-// export static members of this process
+// export static members of this process (lower case)
 export { env } from "./classes/process/env";
 export { actionheroVersion } from "./classes/process/actionheroVersion";
 export { projectRoot } from "./classes/process/projectRoot";
 export { typescript } from "./classes/process/typescript";
 export { id } from "./classes/process/id";
+export {
+  watchFileAndAct,
+  unWatchAllFiles
+} from "./classes/process/watchFileAndAct";
 
 // API object to hold connections, actions, tasks, initializers, and servers
 import { Api } from "./classes/api";
